implementing a common arguments/options parsing

+++ tools/osm/osmdb_buildStructures.py
@@ -14,7 +14,7 @@ and <OUTPUT_TABLE> is defined as:
 """
 # ===========================================================================
 __author__     = "Daniel Krajzewicz"
-__copyright__  = "Copyright 2016-2024, Institute of Transport Research, German Aerospace Center (DLR)"
+__copyright__  = "Copyright 2016-2025, Institute of Transport Research, German Aerospace Center (DLR)"
 __credits__    = ["Daniel Krajzewicz"]
 __license__    = "EPL 2.0"
 __version__    = "0.8.2"
@@ -72,6 +72,8 @@ class OSMExtractor:
         self._objectGeoms = { "node": {}, "way": {}, "rel": {} }
         # seen IDs for removing duplicate IDs
         self._idMapping = { "node": {}, "way": {}, "rel": {} }
+        # map of load definitions to ids
+        self._id2type = { "node": {}, "way": {}, "rel": {} }
     
 
     def load_definitions(self, file_name):
@@ -166,8 +168,6 @@ class OSMExtractor:
                     sd = sd.strip()
                     if sd=="*":
                         oss = self._get_objects(conn, cursor, schema, prefix, subtype, "*", "*", "*")
-                        k = "*"
-                        v = "*"
                     else:
                         for op in ["!=", "!~", "=", "~"]:
                             if sd.find(op)>=0:
@@ -178,6 +178,10 @@ class OSMExtractor:
                         collected = collected.intersection(oss)
                     else:
                         collected = oss
+                    for os in oss:
+                        if os not in self._id2type[subtype]:
+                            self._id2type[subtype][os] = []
+                        self._id2type[subtype][os].append(sd)
                 if len(self._objectIDs[subtype])==0:
                     self._objectIDs[subtype] = collected
                 else:
@@ -290,7 +294,7 @@ class OSMExtractor:
         return area
 
 
-    def _check_commit(self, forced, entries, conn, cursor, schema, name):
+    def _check_commit(self, forced, entries, types, conn, cursor, schema, name):
         """Inserts read objects if forced or if their number is higher than 10000
         :param entries: Descriptions of the objects to insert
         :type entries: list[str]
@@ -311,14 +315,20 @@ class OSMExtractor:
         args = ','.join(cursor.mogrify("(%s, %s, %s, ST_GeomFromText(%s, 4326), ST_GeomFromText(%s, 4326), ST_Centroid(ST_ConvexHull(ST_GeomFromText(%s, 4326))))", i).decode('utf-8') for i in entries)
         cursor.execute("INSERT INTO %s.%s(id, oid, type, polygon, geom_collection, centroid) VALUES " % (schema, name) + (args))
         conn.commit()
+        args = ','.join(cursor.mogrify("(%s, %s, %s)", i).decode('utf-8') for i in types)
+        cursor.execute("INSERT INTO %s.%s_types(id, oid, type) VALUES " % (schema, name) + (args))
+        conn.commit()
         del entries[:]
+        del types[:]
         
 
-    def _add_item(self, entries, item, conn, cursor, schema, name):
+    def _add_item(self, entries, types, item, conn, cursor, schema, name):
         """Prebuilds the given item's insertion string and checks whether it shall be submitted
         :param entries: Descriptions of the objects to insert to extend
         :type entries: list[str]
-        :param item: The item to add to the objects
+        :param types: The types of the objects to add
+        :type types: list[str]
+        :param item: The item to add to the objects and types
         :type item: 
         :param conn: The connection to the database
         :type conn: psycopg2.extensions.connection
@@ -365,7 +375,9 @@ class OSMExtractor:
         else:
             polys = "MULTIPOLYGON EMPTY"
         entries.append([id, oid, type, polys, geom, centroid])
-        self._check_commit(False, entries, conn, cursor, schema, name)
+        for t in self._id2type[type][oid]:
+            types.append([id, oid, t])
+        self._check_commit(False, entries, types, conn, cursor, schema, name)
 
 
     def store_objects(self, area, conn, cursor, schema, name):
@@ -381,20 +393,22 @@ class OSMExtractor:
         :type name: str
         :todo: Make database connection an attribute of the class
         """
-        geometries = []
+        entries = []
+        types = []
         fr = 0
         for rID in self._objectIDs["rel"]:
             if area._relations[rID].build_geometry(area):
-                self._add_item(geometries, area._relations[rID], conn, cursor, schema, name)
+                self._add_item(entries, types, area._relations[rID], conn, cursor, schema, name)
             else: fr += 1
         fw = 0
         for wID in self._objectIDs["way"]:
             if area._ways[wID].build_geometry(area):
-                self._add_item(geometries, area._ways[wID], conn, cursor, schema, name)
+                self._add_item(entries, types, area._ways[wID], conn, cursor, schema, name)
             else: fw += 1
         for nID in self._objectIDs["node"]:
-            self._add_item(geometries, area._nodes[nID], conn, cursor, schema, name)
-        self._check_commit(True, geometries, conn, cursor, schema, name)
+            self._add_item(entries, types, area._nodes[nID], conn, cursor, schema, name)
+        self._check_commit(True, entries, types, conn, cursor, schema, name)
+        #
         return len(self._objectIDs["node"])+len(self._objectIDs["way"])+len(self._objectIDs["rel"]), fw, fr
 
 
@@ -413,7 +427,7 @@ class OSMExtractor:
 
 # --- function definitions --------------------------------------------------
 # --- main
-def main(srcdb, deffile, dstdb):       
+def build_structures(srcdb, deffile, dstdb, dropprevious, append, verbose):
     t1 = datetime.datetime.now()
     # -- open connection
     (host, db, schema_prefix, user, password) = srcdb.split(",")
@@ -438,12 +452,16 @@ def main(srcdb, deffile, dstdb):
     conn2 = psycopg2.connect("dbname='%s' user='%s' host='%s' password='%s'" % (db, user, host, password))
     cursor2 = conn2.cursor()
     # --- build tables
-    cursor2.execute("DROP TABLE IF EXISTS %s.%s" % (schema, name))
-    cursor2.execute("CREATE TABLE %s.%s ( id bigint, oid bigint, type varchar(4) );" % (schema, name))
-    cursor2.execute("SELECT AddGeometryColumn('%s', '%s', 'centroid', 4326, 'POINT', 2);" % (schema, name))
-    cursor2.execute("SELECT AddGeometryColumn('%s', '%s', 'polygon', 4326, 'MULTIPOLYGON', 2);" % (schema, name))
-    cursor2.execute("SELECT AddGeometryColumn('%s', '%s', 'geom_collection', 4326, 'GEOMETRYCOLLECTION', 2);" % (schema, name))
-    conn2.commit()
+    if dropprevious:
+        cursor2.execute("DROP TABLE IF EXISTS %s.%s" % (schema, name))
+    if not append:
+        cursor2.execute("CREATE TABLE %s.%s ( id bigint, oid bigint, type varchar(4) );" % (schema, name))
+        cursor2.execute("SELECT AddGeometryColumn('%s', '%s', 'centroid', 4326, 'POINT', 2);" % (schema, name))
+        cursor2.execute("SELECT AddGeometryColumn('%s', '%s', 'polygon', 4326, 'MULTIPOLYGON', 2);" % (schema, name))
+        cursor2.execute("SELECT AddGeometryColumn('%s', '%s', 'geom_collection', 4326, 'GEOMETRYCOLLECTION', 2);" % (schema, name))
+        cursor2.execute("DROP TABLE IF EXISTS %s.%s_types" % (schema, name))
+        cursor2.execute("CREATE TABLE %s.%s_types (id bigint, oid bigint, type text);" % (schema, name))
+        conn2.commit()
     # --- insert objects
     print ("Building and storing objects")
     num, fw, fr = extractor.store_objects(area, conn2, cursor2, schema, name)
@@ -458,17 +476,68 @@ def main(srcdb, deffile, dstdb):
     if fr>0: print (" %s relations could not be build" % fr)
 
 
+
+# --- function definitions --------------------------------------------------
+# -- main
+def main(arguments=None):
+    """Main method"""
+    # parse options
+    if arguments is None:
+        arguments = sys.argv[1:]
+    # https://stackoverflow.com/questions/3609852/which-is-the-best-way-to-allow-configuration-options-be-overridden-at-the-comman
+    defaults = {}
+    conf_parser = argparse.ArgumentParser(prog='osmdb_buildStructures', add_help=False)
+    conf_parser.add_argument("-c", "--config", metavar="FILE", help="Reads the named configuration file")
+    args, remaining_argv = conf_parser.parse_known_args(arguments)
+    if args.config is not None:
+        if not os.path.exists(args.config):
+            print ("osmdb_buildStructures: error: configuration file '%s' does not exist" % str(args.config), file=sys.stderr)
+            raise SystemExit(2)
+        config = configparser.ConfigParser()
+        config.read([args.config])
+        defaults.update(dict(config.items("DEFAULT")))
+    parser = argparse.ArgumentParser(prog='osmdb_buildStructures', parents=[conf_parser], 
+        description='Builds an road network table using an OSM-database representation', 
+        epilog='(c) Copyright 2016-2025, German Aerospace Center (DLR)')
+    parser.add_argument('OSMdatabase', metavar='OSM-database', help='The definition of the database to read data from;\n'
+            + ' should be a string of the form <HOST>,<DB>,<SCHEMA>.<TABLE_PREFIX>,<USER>,<PASSWD>')
+    parser.add_argument('definition', help='Defines the file to load the definitions of things to extract from')
+    parser.add_argument('output', metavar='OSM-database', help='The definition of the database to read data from;\n'
+            + ' should be a string of the form <HOST>,<DB>,<SCHEMA>.<TABLE_PREFIX>,<USER>,<PASSWD>')
+    parser.add_argument('-R', '--dropprevious', action='store_true', help="Delete destination tables if already existing")
+    parser.add_argument('-A', '--append', action='store_true', help="Append read data to existing tables")
+    parser.add_argument('--version', action='version', version='%(prog)s 0.8.2')
+    parser.add_argument("-v", "--verbose", action="store_true", help="Print what is being done")
+    parser.set_defaults(**defaults)
+    args = parser.parse_args(remaining_argv)
+
+    # check and parse command line parameter and input files
+    errors = []
+    # - input db
+    if len(args.OSMdatabase.split(","))!=5:
+        errors.append("Missing values in target database definition;\n must be: <HOST>,<DB>,<SCHEMA>.<TABLE_PREFIX>,<USER>,<PASSWD>")
+    elif args.OSMdatabase.split(",")[2].count(".")!=1:
+        errors.append("The second field of the target database definition must have the format <SCHEMA>.<TABLE_PREFIX>")
+    # - output db
+    if len(args.output.split(","))!=5:
+        errors.append("Missing values in target database definition;\n must be: <HOST>,<DB>,<SCHEMA>.<TABLE_PREFIX>,<USER>,<PASSWD>")
+    elif args.output.split(",")[2].count(".")!=1:
+        errors.append("The second field of the target database definition must have the format <SCHEMA>.<TABLE_PREFIX>")
+    # - definition file
+    if not os.path.exists(args.definition):
+        errors.append(f"Missing definition file '{args.definition}'")
+    # - report
+    if len(errors)!=0:
+        parser.print_usage(sys.stderr)
+        for e in errors:
+            print ("osmdb_buildStructures: error: %s" % e, file=sys.stderr)
+        print ("osmdb_buildStructures: quitting on error.", file=sys.stderr)
+        return 1
+
+    return build_structures(args.OSMdatabase, args.definition, args.output, args.dropprevious, args.append, args.verbose)
+
+
 # -- main check
 if __name__ == '__main__':
-    if len(sys.argv)<4:
-        print ("""Error: Parameter is missing
-Please run with:
-    osmdb_buildStructures.py <INPUT_TABLES_PREFIX> <DEF_FILE> <OUTPUT_TABLE>
-where <INPUT_TABLES_PREFIX> is defined as:
-    <HOST>,<DB>,<SCHEMA>,<PREFIX>.<USER>,<PASSWD>  
-and <OUTPUT_TABLE> is defined as:
-    <HOST>,<DB>,<SCHEMA>.<NAME>,<USER>,<PASSWD>""")
-        sys.exit(1)
-    main(sys.argv[1], sys.argv[2], sys.argv[3])
-    sys.exit(0)
+    sys.exit(main(sys.argv[1:]))
